% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-constructors.R
\name{AR1}
\alias{AR1}
\title{Specify an Autoregressive Model}
\usage{
AR1(scale_obs = NULL, scale_level = NULL, damp_min = 0.8, damp_max = 0.98)
}
\arguments{
\item{scale_obs}{Scale parameter governing
expected size of observation errors.}

\item{scale_level}{Scale parameter governing
expected size of errors for level term.}

\item{damp_min, damp_max}{Minimum and maximum values
for damping parameter. Defaults are \code{0.8} and
\code{0.98}.}
}
\value{
An object of class \code{"BayesProj_spec_dampedtrend"}.
}
\description{
Specify an order-1 autoregressive time series model.
}
\section{Mathematical details}{


The order-1 autoregressive time series model has the form

\deqn{y_t = \alpha_t + \xi_t}

\deqn{\alpha_t = \mu + \phi (\alpha_{t-1} - \mu) + \epsilon_t}

where
\itemize{
\item \eqn{y_t} is observed  values,
\item \eqn{\alpha_t} is a 'level' term,
\item \eqn{\mu} is a mean term,
\item \eqn{\phi} controls the amount of damping, and
\item \eqn{\xi_t} and \eqn{\epsilon_t} are
normally-distributed 'error' or 'innovation' terms.
}

The mean term \eqn{\mu} has a diffuse normal prior, centered on 0.

The damping parameter \eqn{\phi} is restricted to the interval

\deqn{\phi_{\text{min}} < \phi < \phi_{\text{max}}}

where \eqn{\phi_{\text{min}}} is argument \code{damp_min} and
\eqn{\phi_{\text{max}}} is argument \code{damp_max}.

The standard devations for \eqn{\xi_t},
and \eqn{\epsilon_t} have half normal priors.
Each of these
priors has an associated scale parameter.
}

\examples{
AR1()

## tighter prior on amount of damping
AR1(damp_min = 0.9,
    damp_max = 0.95)
}
