% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-methods.R
\name{augment.BayesProj_proj}
\alias{augment.BayesProj_proj}
\title{Projection of Indicator, Combined with Benchmarks}
\usage{
\method{augment}{BayesProj_proj}(x, interval = 0.95, ...)
}
\arguments{
\item{x}{Results from a call to \code{\link[=project_ts]{project_ts()}}.}

\item{interval}{Width of credible intervals.
A number between 0 and 1. Default is \code{0.95}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble},
containing the 'by' variables from the historical estimates,
the time variable, and the following variables:
\itemize{
\item \code{.fitted}: Point estimates (posterior medians) of rates.
\item \code{.lower}, \code{.upper}: Lower and upper bounds of
credible intervals specified by \code{interval}.
\item \code{.probability}: A list column with all draws
from the posterior distribution.
}

If a value for \code{spec_bench} was provided in the call
to \code{\link[=project_ts]{project_ts()}}, then information on the benchmarks
is also included.
}
\description{
Given output from a call to \code{\link[=project_ts]{project_ts()}},
extract forecasts for the demographic indicator,
merged with the benchmarks (if used).
}
\details{
The values for the indicator, and for the benchmarks
are shown in the original units, even if
\code{log = TRUE} in the original call to
\code{\link[=fit_ts]{fit_ts()}}.
}
\examples{
library(tibble)
data <- tribble(~time, ~val,
                2015,  4.3,
                2016,  4.1,
                2020,  3.5,
                2022,  3.6)
fitted <- fit_ts(data, indvar = "val")
bench <- Benchmarks(tribble(~time, ~q50, ~q90,
                             2030,  1.8,  2.3))
projected <- project_ts(fitted = fitted,
                        time_labels = 2023:2030,
                        spec_bench = bench)
augment(projected)
}
\seealso{
\itemize{
\item \code{\link[=fit_ts]{fit_ts()}} to fit a time series model to historical data.
\item \code{\link[=project_ts]{project_ts()}} to forecast a time series model.
\item \code{\link[=Benchmarks]{Benchmarks()}} to specify "benchmarks" (expert judgements
about plausible future values) for a forcast.
\item \code{\link[=components.BayesProj_proj]{components()}}
extracts forecasts for parameters and hyper-parameter.
\item \code{\link[=n_draw<-]{n_draw()}} sets the default number
of draws from the posterior distribution.
\item \code{\link[=composite]{composite()}} combines estimates and projections
in a single tibble.
}
}
