% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-functions.R
\name{composite}
\alias{composite}
\title{Combine Results from Estimates and Projection}
\usage{
composite(x, interval = 0.95)
}
\arguments{
\item{x}{Results from a call to \code{\link[=project_ts]{project_ts()}}.}

\item{interval}{Width of credible intervals.
A number between 0 and 1. Default is \code{0.95}.}
}
\value{
A tibble
}
\description{
Combine estimated and projected values for demographic
indicator, in a format suitable for graphing
or analysis.
}
\examples{
library(tibble)
data <- tribble(~time, ~val,
                2015,  4.3,
                2016,  4.1,
                2020,  3.5,
                2022,  3.6)
fitted <- fit_ts(data, indvar = "val")
bench <- Benchmarks(tribble(~time, ~q50, ~q90,
                            2030,  1.8,  2.3))
projected <- project_ts(fitted = fitted,
                        time_labels = 2023:2030,
                        spec_bench = bench)
composite(projected)
}
\seealso{
\itemize{
\item \code{\link[=project_ts]{project_ts()}} to project a time series model
\item \code{\link[=augment.BayesProj_proj]{augment()}}
combines data and benchmarks.
\item \code{\link[=components.BayesProj_proj]{components()}}
extracts forecasts for parameters and hyper-parameter.
}
}
