% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-constructors.R
\name{DampedTrend}
\alias{DampedTrend}
\title{Specify a Damped Linear Trend Model}
\usage{
DampedTrend(
  scale_obs = NULL,
  scale_level = NULL,
  scale_trend = NULL,
  damp_min = 0.8,
  damp_max = 0.98
)
}
\arguments{
\item{scale_obs}{Scale parameter governing
expected size of observation errors.}

\item{scale_level}{Scale parameter governing
expected size of errors for level term.}

\item{scale_trend}{Scale parameter governing
expected size of errors for trend term.}

\item{damp_min, damp_max}{Minimum and maximum values
for damping parameter. Defaults are \code{0.8} and
\code{0.98}.}
}
\value{
An object of class \code{"BayesProj_spec_dampedtrend"}.
}
\description{
Specify a "damped linear trend" time series model.
}
\section{Mathematical details}{


The damped linear trend time series model has the form

\deqn{y_t = \alpha_t + \xi_t}

\deqn{\alpha_t = \alpha_{t-1} + \delta_{t-1} + \epsilon_t}

\deqn{\delta_t = \phi \delta_{t-1} + \varepsilon_t}

where
\itemize{
\item \eqn{y_t} is observed  values,
\item \eqn{\alpha_t} is a 'level' term,
\item \eqn{\delta_t} is a 'trend' term,
\item \eqn{\phi} controls the amount of damping, and
\item \eqn{\xi_t}, \eqn{\epsilon_t}, and \eqn{\varepsilon_} are all
normally-distributed 'error' or 'innovation' terms.
}

The damping parameter \eqn{\phi} is restricted to the interval

\deqn{\phi_{\text{min}} < \phi < \phi_{\text{max}}}

where \eqn{\phi_{\text{min}}} is argument \code{damp_min} and
\eqn{\phi_{\text{max}}} is argument \code{damp_max}.

The standard devations for \eqn{\xi_t},
\eqn{\epsilon_t}, and \eqn{\varepsilon}
have half normal priors. Each of these
priors has an associated scale parameter.
}

\examples{
DampedTrend()

## allow for more rapid change
DampedTrend(scale_level = 0.5,
            scale_trend = 0.5)

## tighter prior on amount of damping
DampedTrend(damp_min = 0.9,
            damp_max = 0.95)
}
