% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ts.R
\name{fit_ts}
\alias{fit_ts}
\title{Fit a Time Series Model}
\usage{
fit_ts(
  data,
  indvar,
  timevar = "time",
  byvar = character(),
  spec_ts = DampedTrend(),
  log = FALSE
)
}
\arguments{
\item{data}{Data frame containing estimates
for summary indicator.}

\item{indvar}{Name of the variable holding
the demographic indicator.}

\item{timevar}{Name of the time variable.}

\item{byvar}{Names of classification variables
in \code{data}. Optional.}

\item{spec_ts}{Specification for a time series model.
See \code{\link[=DampedTrend]{DampedTrend()}}.}

\item{log}{Whether to estimate the model on the
log scale. Default is \code{FALSE}.}
}
\value{
An object of class \code{"BayesProj_fitted"}.
}
\description{
Use historical data to obtain estimates
for the parameters in a time series model.
Typically used to format a
demographic indicator, such as the
total fertility rate.
}
\section{Format of indicator}{


The demographic indicator can be a single set of estimates
or it can be multiple draws from a posterior
distribution. If it is a single set of estimates,
then \code{indvar} should be an ordinary numeric vector.
If it is multiple draws, then it should be a
list of vectors, each of which has the same length.
For an example of multiple draws, see \code{\link[BayesRates:total_rate]{BayesRates::total_rate()}}.
}

\section{'by' variable}{


\code{model_ts()} has an optional \code{byvar} argument,
specifying variables in \code{data}. If
a value for \code{byvar} is supplied, then
separate models are fitted for each combination
of the 'by' variables.
}

\section{Modelling on the log scale}{


If the \code{log} argument is \code{TRUE}, then the
model is fitted to logged values of the
indicator. This can be appropriate when an indicator
is close to zero and has more room to move
up than down.
}

\seealso{
\itemize{
\item \code{\link[=DampedTrend]{DampedTrend()}} to specify a "damped linear trend"
time series model.
\item \link{project_ts} to project an indicator, using
results from a call to \code{model_ts()}.
\item \code{\link[BayesRates:total_rate]{BayesRates::total_rate()}} to calculate a demographic
indicator.
\item \code{\link[=components]{components()}} for parameters from the fitted model
\item \code{\link[=augment]{augment()}} for smoothed values of the indicator
from the fitted model, combined with the original data.
}
}
