% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-methods.R
\name{make_labels_fit}
\alias{make_labels_fit}
\title{Derive Labels to Use When Reporting on Fitted Model}
\usage{
make_labels_fit(spec, labels_time)
}
\arguments{
\item{spec}{Object of class 'BayesProj_spec_ts'}

\item{labels_time}{Vector of labels for historical periods,
including periods with no data}
}
\value{
A character vector

noRd
}
\description{
Creates labels for a single combination of
the 'by' variables
}
