% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-methods.R
\name{make_labels_proj}
\alias{make_labels_proj}
\title{Derive Labels to Use When Reporting on Projection}
\usage{
make_labels_proj(spec, labels_time_project)
}
\arguments{
\item{spec}{Object of class 'BayesProj_spec_ts'}

\item{labels_time_project}{Vector of labels for projection periods}
}
\value{
A character vector

noRd
}
\description{
Creates labels for a single combination of
the 'by' variables
}
