% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-methods.R
\name{make_vname_proj}
\alias{make_vname_proj}
\title{Derive Names of Variables to Use When Reporting on Projection}
\usage{
make_vname_proj(spec, timevar)
}
\arguments{
\item{spec}{Object of class 'BayesProj_spec_ts'}

\item{timevar}{Name of the time variable}
}
\value{
A character vector

noRd
}
\description{
Creates identifiers for a single combination of
the 'by' variables
}
