% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ts-methods.R
\name{make_what_proj}
\alias{make_what_proj}
\title{Derive Identifiers for Batches of Parameters
to Use When Reporting on Projection}
\usage{
make_what_proj(spec, indvar, labels_time_project)
}
\arguments{
\item{spec}{Object of class 'BayesProj_spec_ts'}

\item{indvar}{Name of indicator variable}

\item{labels_time_project}{Vector of labels for
projection periods}
}
\value{
A character vector

noRd
}
\description{
Creates identifiers for a single combination of
the 'by' variables
}
