% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-methods.R
\name{n_draw<-}
\alias{n_draw<-}
\title{Set Number of Draws from Posterior Distribution}
\usage{
n_draw(x) <- value
}
\arguments{
\item{x}{An object of class \code{"BayesProj_fitted"},
generated by a call to \code{\link[=fit_ts]{fit_ts()}}.}

\item{value}{Number of posterior draws. A non-negative
integer.}
}
\value{
A modified version of \code{x}.
}
\description{
Specify the number of draws from the posterior
distribution that are generated by a results object from
\code{\link[=fit_ts]{fit_ts()}}.
}
\examples{
library(tibble)
data <- tribble(~time, ~val,
                2015,  4.3,
                2016,  4.1,
                2020,  3.5,
                2022,  3.6)
fit <- fit_ts(data, indvar = "val")
fit
n_draw(fit) <- 10
fit
}
\seealso{
\itemize{
\item \code{\link[=fit_ts]{fit_ts()}} fits a time series model to historical data.
\item \code{\link[=augment.BayesProj_fitted]{augment()}}
combines data and estimates.
\item \code{\link[=components.BayesProj_fitted]{components()}}
extracts estimates of parameters and hyper-parameters.
}
}
