% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ts.R
\name{project_ts}
\alias{project_ts}
\title{Project a Demographic Indicator}
\usage{
project_ts(fitted, time_labels, spec_bench = NULL)
}
\arguments{
\item{fitted}{A fitted time series model,
with class \code{"BayesProj_fitted"}. Typically
the results from a call to \code{\link[=fit_ts]{fit_ts()}}.}

\item{time_labels}{Character vector of time labels.}

\item{spec_bench}{Specification of forecast
benchmarks. Typically the results from a call
to \code{\link[=Benchmarks]{Benchmarks()}}. Optional.}
}
\value{
An object of class \code{"BayesProj_proj"}.
}
\description{
Construct a projection for a
demographic indicator, based on a fitted
model, and, optionally, on expert judgement.
The expert judgement is expressed through
benchmarks for future values.
}
\examples{
library(tibble)
data <- tribble(~time, ~val,
                2015,  4.3,
                2016,  4.1,
                2020,  3.5,
                2022,  3.6)
fitted <- fit_ts(data, indvar = "val")
bench <- tribble(~time, ~q50, ~q90,
                 2030,  1.8,  2.3)
projected <- project_ts(fitted = fitted,
                        time_labels = 2023:2030,
                        spec_bench = bench)
projected
}
\seealso{
\itemize{
\item \code{\link[=fit_ts]{fit_ts()}} to fit a time series model.
\item \code{\link[=Benchmarks]{Benchmarks()}} to specify forecast benchmarks.
\item \code{\link[=components]{components()}} for parameters from the projection.
\item \code{\link[=augment]{augment()}} for projections of the indicator,
together with benchmarks (if supplied).
}
}
