% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesProj-package.R
\docType{package}
\name{BayesProj-package}
\alias{BayesProj}
\alias{BayesProj-package}
\title{Project Summary Indicators for Demographic Rates
using Bayesian Methods}
\description{
Use Bayesian methods to project demographic rates.
The projections typically involve some expert
judgement, expressed through benchmarks.
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Junni Zhang \email{junnizhang@163.com}
}

}
\keyword{internal}
