% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-methods.R
\name{components.BayesProj_fitted}
\alias{components.BayesProj_fitted}
\title{Extract Components from a Fitted Time Series Model}
\usage{
\method{components}{BayesProj_fitted}(object, what = NULL, interval = 0.95, ...)
}
\arguments{
\item{object}{Results from a call to \code{\link[=fit_ts]{fit_ts()}}.}

\item{what}{Components to be extracted.
If no value supplied, all available components
are extracted.}

\item{interval}{Width of credible intervals.
A number between 0 and 1. Default is \code{0.95}.}

\item{...}{Not currently used.}
}
\value{
A named list of tibbles, or, if \code{what}
is a single value, a single tibble.
Each tibble contains
\itemize{
\item classification variables, if present
\item \code{.fitted} Point estimates (posterior medians).
\item \code{.lower}, \code{.upper} Lower and upper bounds of
credible intervals.
\item \code{.probability}. A list column with all draws
from the posterior distribution.
}
}
\description{
Extract Components from a Fitted Time Series Model
}
\examples{
library(tibble)
data <- tribble(~time, ~val,
                2015,  4.3,
                2016,  4.1,
                2020,  3.5,
                2022,  3.6)
fit <- fit_ts(data, indvar = "val")
components(fit)
}
\seealso{
\itemize{
\item \code{\link[=fit_ts]{fit_ts()}} to fit a time series model
\item \code{\link[=augment.BayesProj_fitted]{augment()}}
combines data and estimates.
\item \code{\link[=n_draw<-]{n_draw()}} sets the default number
of draws from the posterior distribution.
}
}
