% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ts.R
\name{project_ts}
\alias{project_ts}
\title{Project a Demographic Indicator}
\usage{
project_ts(mod, spec_bench = NULL)
}
\arguments{
\item{mod}{A fitted time series model,
with class \code{"BayesProj_fitted"}. Typically
the results from a call to \code{\link[=fit_mod]{fit_mod()}}.}

\item{spec_bench}{Specification of forecast
benchmarks. Typically the results from a call
to \code{\link[=Benchmarks]{Benchmarks()}}. Optional.}
}
\value{
An object of class \code{"BayesProj_proj"}.
}
\description{
Project a demographic indicator, based on a fitted
model, and, optionally, on expert judgement.
The expert judgement is expressed through
benchmarks for future values.
}
\seealso{
\itemize{
\item \code{\link[=fit_mod]{fit_mod()}} to fit a time series model.
\item \code{\link[=Benchmarks]{Benchmarks()}} to specify forecast benchmarks.
time series model.
\item \code{\link[=components]{components()}} for parameters from the projection.
\item \code{\link[=augment]{augment()}} for future values from the the projection,
combined with the benchmarks.
}
}
